# Improve performance of class variables