--- a/src/main/java/org/truffleruby/core/module/ModuleFields.java
+++ b/src/main/java/org/truffleruby/core/module/ModuleFields.java
@@ -21,6 +21,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.locks.ReentrantLock;
 
+import com.oracle.truffle.api.object.DynamicObjectLibrary;
 import org.jcodings.specific.UTF8Encoding;
 import org.truffleruby.RubyContext;
 import org.truffleruby.RubyLanguage;
@@ -46,6 +47,7 @@ import org.truffleruby.language.loader.ReentrantLockFreeingMap;
 import org.truffleruby.language.methods.InternalMethod;
 import org.truffleruby.language.objects.ObjectGraph;
 import org.truffleruby.language.objects.ObjectGraphNode;
+import org.truffleruby.language.objects.classvariables.ClassVariableStorage;
 import org.truffleruby.language.objects.shared.SharedObjects;
 
 import com.oracle.truffle.api.Assumption;
@@ -95,7 +97,7 @@ public class ModuleFields extends ModuleChain implements ObjectGraphNode {
 
     private final ConcurrentMap<String, InternalMethod> methods = new ConcurrentHashMap<>();
     private final ConcurrentMap<String, RubyConstant> constants = new ConcurrentHashMap<>();
-    private final ConcurrentMap<String, Object> classVariables = new ConcurrentHashMap<>();
+    private final ClassVariableStorage classVariables = new ClassVariableStorage();
 
     /** The refinements (calls to Module#refine) nested under/contained in this namespace module (M). Represented as a
      * map of refined classes and modules (C) to refinement modules (R). */
@@ -193,7 +195,9 @@ public class ModuleFields extends ModuleChain implements ObjectGraphNode {
             this.constants.put(entry.getKey(), entry.getValue());
         }
 
-        this.classVariables.putAll(fromFields.classVariables);
+        for (Object key : fromFields.classVariables.getShape().getKeys()) {
+            DynamicObjectLibrary.getUncached().put(this.classVariables, key, DynamicObjectLibrary.getUncached().getOrDefault(fromFields.classVariables, key, null));
+        }
 
         if (fromFields.hasPrependedModules()) {
             // Then the parent is the first in the prepend chain
@@ -745,7 +749,7 @@ public class ModuleFields extends ModuleChain implements ObjectGraphNode {
         return methods.get(name);
     }
 
-    public ConcurrentMap<String, Object> getClassVariables() {
+    public ClassVariableStorage getClassVariables() {
         return classVariables;
     }
 
@@ -849,8 +853,8 @@ public class ModuleFields extends ModuleChain implements ObjectGraphNode {
             ObjectGraph.addProperty(adjacent, constant);
         }
 
-        for (Object value : classVariables.values()) {
-            ObjectGraph.addProperty(adjacent, value);
+        for (Object key : classVariables.getShape().getKeys()) {
+            ObjectGraph.addProperty(adjacent, DynamicObjectLibrary.getUncached().getOrDefault(classVariables, key, null));
         }
 
         for (InternalMethod method : methods.values()) {