… CHANGELOG.md
… src/main/java/org/truffleruby/core/module/RubyModule.java
… src/main/java/org/truffleruby/language/LexicalScope.java
… src/main/java/org/truffleruby/language/objects/classvariables/CheckClassVariableNameNode.java
… src/main/java/org/truffleruby/language/objects/classvariables/ClassVariableStorage.java
… src/main/java/org/truffleruby/language/objects/classvariables/LookupClassVariableNode.java
… src/main/java/org/truffleruby/language/objects/classvariables/LookupClassVariableStorageNode.java
… src/main/java/org/truffleruby/language/objects/classvariables/ReadClassVariableNode.java
… src/main/java/org/truffleruby/language/objects/classvariables/ResolveTargetModuleForClassVariablesNode.java
… src/main/java/org/truffleruby/language/objects/classvariables/SetClassVariableNode.java
… src/main/java/org/truffleruby/language/objects/classvariables/WriteClassVariableNode.java
… src/main/java/org/truffleruby/parser/BodyTranslator.java