Someone asked me how TruffleRuby optimised class variables. I went to check and was surprised to find it completely unoptimised. By pulling out a few fast-paths from the code behind boundaries and switching the data structure used to store the class variables I've managed to significantly improve performance for a few common cases.

Before:

```
           ivar_read    455.654M (±15.7%) i/s -      2.238B in   5.090222s
                read     48.886M (± 9.9%) i/s -    243.582M in   5.050903s
          class_read     57.385M (± 7.5%) i/s -    286.451M in   5.034053s
       constant_read     47.889M (± 5.2%) i/s -    239.445M in   5.016469s
               write     13.938M (± 4.6%) i/s -     70.387M in   5.061929s
         class_write     14.790M (± 6.1%) i/s -     74.568M in   5.063688s
                 get     30.744M (± 8.0%) i/s -    154.151M in   5.050587s
                 set     11.971M (± 5.4%) i/s -     59.695M in   5.003366s
          defined_op     49.961M (± 2.3%) i/s -    249.753M in   5.001779s
      defined_method     31.913M (± 6.0%) i/s -    159.384M in   5.014710s
     class_variables      5.048M (± 6.2%) i/s -     25.422M in   5.057317s
```

After:

```
           ivar_read    493.733M (± 9.3%) i/s -      2.345B in   4.963858s
                read    533.243M (±16.8%) i/s -      2.583B in   5.101963s
          class_read    796.643M (± 9.3%) i/s -      4.077B in   5.165160s
       constant_read    443.462M (±14.0%) i/s -      2.168B in   5.058631s
               write    654.199M (± 1.7%) i/s -      3.305B in   5.053695s
         class_write    977.600M (± 2.5%) i/s -      4.929B in   5.045495s
                 get    560.218M (± 2.9%) i/s -      2.818B in   5.034547s
                 set    355.147M (± 3.7%) i/s -      1.795B in   5.061929s
          defined_op    976.669M (± 3.2%) i/s -      4.937B in   5.060990s
      defined_method    560.041M (± 2.6%) i/s -      2.834B in   5.064155s
     class_variables      4.194M (± 1.7%) i/s -     21.012M in   5.011955s
```

Difference:

```
           ivar_read    no change
                read    > 10x
          class_read    > 10x
       constant_read    10x
               write    50x
         class_write    > 50x
                 get    20x
                 set    30x
          defined_op    20x
      defined_method    20x
     class_variables    a little slower
```

Compared to MRI:

```
ruby 2.7.2p137 (2020-10-01 revision 5445e04352) [x86_64-darwin19]
           ivar_read     18.903M (± 1.9%) i/s -     95.796M in   5.069657s
                read     12.228M (± 2.2%) i/s -     61.593M in   5.039576s
          class_read     11.812M (± 3.1%) i/s -     59.006M in   5.001077s
       constant_read     12.110M (± 2.1%) i/s -     61.383M in   5.071081s
               write      5.226M (± 2.1%) i/s -     26.613M in   5.094473s
                 get      8.520M (± 1.7%) i/s -     43.100M in   5.059971s
                 set      4.328M (± 3.2%) i/s -     21.725M in   5.025026s
          defined_op     14.676M (± 1.9%) i/s -     74.354M in   5.068288s
      defined_method     10.210M (± 3.6%) i/s -     51.086M in   5.010694s
     class_variables      2.055M (± 2.3%) i/s -     10.320M in   5.023387s
```

Compared to JRuby:

```
jruby 9.2.14.0 (2.5.7) 2020-12-08 ebe64bafb9 OpenJDK 64-Bit Server VM 25.252-b14 on 1.8.0_252-b14 +jit [darwin-x86_64]
           ivar_read     23.438M (±10.7%) i/s -    113.363M in   5.001581s
                read     16.011M (± 2.8%) i/s -     80.148M in   5.009717s
          class_read     17.961M (± 7.4%) i/s -     89.175M in   4.998767s
       constant_read     16.231M (± 5.7%) i/s -     80.991M in   5.008429s
               write      8.229M (± 3.1%) i/s -     41.137M in   5.004191s
                 get      8.100M (± 2.0%) i/s -     40.694M in   5.025950s
                 set      5.521M (± 7.6%) i/s -     27.697M in   5.059839s
          defined_op     23.834M (± 8.7%) i/s -    116.812M in   5.011901s
      defined_method     10.295M (± 4.2%) i/s -     51.524M in   5.014946s
     class_variables      3.978M (± 3.8%) i/s -     20.043M in   5.046777s
```

What have I done?

* switched the data structure from a hash map to a dynamic object (an object with a hidden class), which allows me to inline cache the location of it within the object and avoid boxing when storing a variable and unboxing when reading it
* split operations on class variables into separate operations to find the right module for the variable, and then for reading the variable from that module
* identified a few likely common fast paths that can be checked with guards

What does it look like?

For example, a simple class variable read in an instance method.

```ruby
class Foo
  @@example = 14
  def example_read
    @@example
  end
end
```

This ends up as this graph. The class variable storage object can be turned into a constant, coming from the method's lexical scope at the point it was defined. We can then read the shape of the class variable storage and check it against an expected shape using an identity comparison. If that fails, we deoptimise. We can then cache the location within the storage of the class variable, so that becomes nothing more than a field read.

![example_read](https://user-images.githubusercontent.com/341094/108004124-cd2a5280-6fec-11eb-86b2-cebadb9665ee.png)

So a class variable read from an instance method of the class defining the class variable can be as simple as:

```s
	0x12347d8f0:	movabs	rax, 0x61095e498
	0x12347d900:	cmp	dword ptr [rax + 0xc], 0xc212bc9b
	0x12347d907:	jne	0x12347d9c3
	0x12347d90d:	mov	r10d, dword ptr [rax + 0x18]
```

What else could we do:

If we find interesting cases not covered by these fast paths, such as particular patterns of using class variables from superclasses or extended modules, we can probably add those for up to a couple of levels of indirection fairly easily.

Benchmark:

```ruby
require 'benchmark/ips'

class Foo

  @@example = 14
  @@example_constant = 14

  def initialize
    @example = 14
  end

  def example_ivar_read
    @example
  end

  def example_read
    @@example
  end

  def self.example_class_read
    @@example
  end

  def example_constant_read
    @@example_constant
  end

  def example_write(value)
    @@example = value
  end

  def self.example_class_write(value)
    @@example = value
  end

  def example_get
    self.class.class_variable_get(:@@example)
  end

  def example_set(value)
    self.class.class_variable_set(:@@example, value)
  end

  def example_defined_op
    defined?(@@example)
  end

  def example_defined_method
    self.class.class_variable_defined?(:@@example)
  end

  def example_class_variables
    self.class.class_variables
  end

end

foo = Foo.new
x = 100

Benchmark.ips do |x|
  x.warmup = 5
  x.time = 5

  x.report('ivar_read') { foo.example_ivar_read }
  x.report('read') { foo.example_read }
  x.report('class_read') { Foo.example_class_read }
  x.report('constant_read') { foo.example_constant_read }
  x.report('write') { foo.example_write(x) }
  x.report('class_write') { Foo.example_class_write(x) }
  x.report('get') { foo.example_get }
  x.report('set') { foo.example_set(x) }
  x.report('defined_op') { foo.example_defined_op }
  x.report('defined_method') { foo.example_defined_method }
  x.report('class_variables') { foo.example_class_variables }
end

```