@chrisseaton 2021-02-16
> Another possibility here might be to have a Map<String name, ClassVariableStorage storage> (and ClassVariableStorage would be a class with a single volatile field value), and then look that one up once, and then directly read/write the storage object (similar to GlobalVariables/GlobalVariableStorage but simpler as there are no hooks).

I thought about that but if we wanted to avoid boxing we'd have to re-invent some of what `DynamicObject` does for storage locations.

> Using a DynamicObject also loses the volatile behavior of class variables, which is described in the proposal for a Ruby memory model.

Is there a barrier we can insert after writes or around reads to get volatile back?