@eregon 2021-02-16
> I thought about that but if we wanted to avoid boxing we'd have to re-invent some of what `DynamicObject` does for storage locations.

Good point.

> Is there a barrier we can insert after writes or around reads to get volatile back?

This might be useful:
https://github.com/graalvm/graal-jvmci-8/blob/210ecc800424aa07f4e03498efff4f4eeb9b6118/jvmci/jdk.vm.ci.code/src/jdk/vm/ci/code/MemoryBarriers.java#L110-L113

And we have a mapping of Unsafe fences here:
https://github.com/AdoptOpenJDK/openjdk-jdk11/blob/19fb8f93c59dfd791f62d41f332db9e306bc1422/src/java.base/share/classes/jdk/internal/misc/Unsafe.java#L3275-L3315

However, I'm not sure modelling volatile accesses with fences is quite correct (e.g., https://shipilev.net/blog/2016/close-encounters-of-jmm-kind/#myth-barriers-are-sane). But probably our best option in this case to still avoid boxing.