@chrisseaton 2021-02-16
The object model uses `Unsafe` to read and write storage locations. Is there a volatile equivalent that the object model could use for us? And actually how does the use of `Unsafe` match up with scalar replacement of aggregates and other optimisations? How does it stop it being bypassed as an uncontrolled unsafe action?