@eregon 2021-02-16
> Is there a volatile equivalent that the object model could use for us?

Yes, there is `put*Volatile`. But AFAIK there is no way to make the object model use that currently. cc @woess 

> And actually how does the use of `Unsafe` match up with scalar replacement of aggregates and other optimisations? How does it stop it being bypassed as an uncontrolled unsafe action?

If the DynamicObject is never allocated then it doesn't matter, nothing else can access those fields anyway.
If it's EA'd and then later on allocated, there is some STORE_STORE barrier, so anyone reading from that object should correctly see all initial values ([safe initialization](https://shipilev.net/blog/2014/safe-public-construction/#_safe_initialization)).