@chrisseaton 2021-02-16
Graal conicalises the `RawRead` back to a `LoadField` - that's what I was confused about.