@eregon 2021-02-16
We have another issue: class variables might be mutated/accessed concurrently.
Before, this was handled by the ConcurrentHashMap, but DynamicObject alone is not thread-safe.

We should synchronize writes, whenever the RubyModule is shared.
In practice, most modules/classes end up being shared between threads (at least when there is a second Ruby Thread) because they are referenced by constants.
So when the RubyModule gets shared, so should the ClassVariableStorage, and this can happen simply in `ModuleFields#getAdjacentObjects`.

We also need write barriers like it's done in ModuleOperations#setClassVariable. PropagateSharingNode can be used for that conveniently on the fast path.

And all write accesses need to either use `WriteObjectFieldNode` or `synchronized` if ClassVariableStorage is shared.